% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{AFDP}
\alias{AFDP}
\title{AFDP data
contains the 11 sensor measures
aggregated over one hour (by means of average or sum) from a gas
turbine located in Turkey’s north western region for the purpose of studying
flue gas emissions
Variable description:}
\description{
AFDP data
contains the 11 sensor measures
aggregated over one hour (by means of average or sum) from a gas
turbine located in Turkey’s north western region for the purpose of studying
flue gas emissions
Variable description:
}
\references{
Chang, G. T., & Walther, G. (2007). Clustering with mixtures of log-concave distributions. Computational Statistics & Data Analysis, 51(12), 6242-6251.
}
\keyword{data}
