% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{distlatfrequency}
\alias{distlatfrequency}
\title{Euclidean Distance Based Labeling Method for Label Switching}
\usage{
distlatfrequency(est, lat)
}
\arguments{
\item{est}{Records the MLE and the estimated classification probabilities, can
be derived by the function: est=mixnorm(x,2,ini).}

\item{lat}{The latent component labels for all observations, a m by n 0-1 matrix
lat if (i,j)th cell is 1, then jth observation is from ith component.}
}
\value{
The estimation by complete posterior distribution
the result vector consist of mu of first dimension of first component, second component...,
mu of second dimension of first component, second component ...
pi of first component , second component ...
}
\description{
Solving the label switching problem  by minimizing the distance between the
classification probabilities and the true latent labels.
}
\examples{
#See examples in complhfrequency function.
}
