% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{elbow}
\alias{elbow}
\title{Elbow data
This dataset contains elbow dimension measurements on 507 individuals. The 247 men and 260
women were primarily individuals in their twenties and thirties, with a
scattering of older men and women, all exercising several hours a week.
Variable description:
Elbow: elbow diameter, sum of two elbows}
\description{
Elbow data
This dataset contains elbow dimension measurements on 507 individuals. The 247 men and 260
women were primarily individuals in their twenties and thirties, with a
scattering of older men and women, all exercising several hours a week.
Variable description:
Elbow: elbow diameter, sum of two elbows
}
\references{
Heinz, G., Peterson, L. J., Johnson, R. W., & Kerk, C. J. (2003). Exploring relationships in body dimensions. Journal of Statistics Education, 11(2)
}
\keyword{data}
