% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethanol}
\alias{ethanol}
\title{Ethanol data}
\format{
A data frame containing 88 observations and the following 3 variables.
\describe{
\item{NO:}{ concentration of nitric oxide (NO) and nitrogen dioxide (NO2) in engine exhaust.}
\item{Compression:}{ compression ratio of engine.}
\item{Equivalence:}{ equivalence ratio. This is a measure of the richness of the air and ethanol fuel mixture.}
}
}
\source{
Original source:

Brinkman, N. D. (1981). Ethanol fuel—single—cylinder engine study of efficiency and exhaust emissions.
SAE transactions, 1410-1424.

R source:

Wand M (2018). SemiPar: Semiparametic Regression. R package version 1.0-4.2,
\url{https://CRAN.R-project.org/package=SemiPar}.

Sarkar D (2008). Lattice: Multivariate Data Visualization with R. Springer, New York. ISBN
978-0-387-75968-5, \url{http://lmdvr.r-forge.r-project.org}.
}
\usage{
ethanol
}
\description{
The data contains 88 sets of measurements of peak nitrogen oxide emission levels from an experiment
in which ethanol was burned in a single-cylinder engine. The emission levels of nitrogen oxides were
recorded under different settings of the compression ratio and equivalence ratio of the engine.
}
\references{
Ruppert, D., Wand, M. P., and Carroll, R. J. (2003). Semiparametric regression (No. 12).
Cambridge university press.

Hurn, M., Justel, A., and Robert, C. P. (2003). Estimating mixtures of regressions.
Journal of computational and graphical statistics, 12(1), 55-79.
}
\keyword{datasets}
