% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varyprop.R
\name{gen_mixreg1}
\alias{gen_mixreg1}
\title{Varpop function Data Generator}
\usage{
gen_mixreg1(n, numc)
}
\arguments{
\item{n}{number of observations}

\item{numc}{number of components}
}
\value{
list of x,y, true mean functions and true p.
}
\description{
Used to generate data for MixReg_Pvary function examples.
}
\examples{
gen_mixreg1(n=100,numc=2)
}
