% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDEEM.R
\name{KDEEM.LSE}
\alias{KDEEM.LSE}
\title{Kernel Density-based EM type algorithm with Least Square Estimation for Semiparametric Mixture Regression with Unspecified Homogenous Error Distributions}
\usage{
KDEEM.LSE(x, y, M, ini)
}
\arguments{
\item{x}{explanatory variables in matrix from with row as observations}

\item{y}{response variable}

\item{M}{number of component}

\item{ini}{initial values for parameters. ini = list(beta = regression coefficients,
posterior = matrix: probability of each observation is from component m)
the initial value can be obtained by applying traditional MLE with EM algorithm and normal assumption.}
}
\value{
P: posterior probability of each observation is from component m
Beta: estimated regression coefficients
Tao: Estimated precision parameter, the inverse of standard deviation.
PI: estimated component proportion
H: bandwidth of the kernel estimation
}
\description{
For kernel density-based semiparametric mixture regression for homogeneous error distributions
Using weighted least square estimation for beta in M step, where weight is the posterior probability of observation comes from each component.
And Gaussian kernel for g(.) component estimation
}
\examples{
# See example in KDEEM
}
