% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixregression.R
\name{mixlinrb_bione}
\alias{mixlinrb_bione}
\title{Robust EM algorithm to fit the mixture of linear regression based on bisquare function with one initial value}
\usage{
mixlinrb_bione(x, y, bet, sig, pr, m = 2)
}
\arguments{
\item{x}{explanatory variables matrix with rows for each observation}

\item{y}{response variable vector}

\item{bet}{initial value of beta}

\item{sig}{initial value of sigma}

\item{pr}{initial value of proportion}

\item{m}{number of components, default is 2}
}
\value{
theta: estimated parameters matrix, the columns are beta(intercept, slopes), sigma, proportion for each component.
run:number of iterations to converge
}
\description{
mixlinrb_bione estimates the mixture regression parameters robustly using bisquare function based on one initial value
}
