% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{mixmnorm}
\alias{mixmnorm}
\title{Parameter Estimation for Multivariate Normal Mixture}
\usage{
mixmnorm(x, sigma_k = NULL, equal = NULL, ini = NULL)
}
\arguments{
\item{x}{observations.}

\item{sigma_k}{Number of component. See details.}

\item{equal}{If the covariate matrix of each component is equal. See details.}

\item{ini}{initial value of parameters}
}
\value{
mu: estimated mean
sigma: estimated sd
pi:estimated proportion
lh: likelihood of estimated parameters
p: density of observations
}
\description{
Estimate parameters of multivariate normal mixtures using EM-algorithm.
}
\details{
equal:possible value: 0,1,.... When equal = 0, fit k component normal mixture with unknown covariance matrix; when equal = 1,
fit k component normal mixture with unknown equal covariance matrix; when equal = other value, fit k component normal mixture with covariance matrix is known.

sigma_k: Default is NULL. When both of sigma_k and equal is NULL then estimate the data as two component
multivariate normal with equal unknown covariates.
}
