% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{NBA}
\alias{NBA}
\title{NBA data
The data set contains four descriptive statistics for all 105 guards for the 1992–1993 season.
There are many ways to measure the (statistical) performance of guards in the NBA. Of interest is how the
height of the player, minutes per game and free throw percentage affect points per game.
Variable description:
Height: height of the player
MPG: minutes per game
FTP: free throw percentage
PPM: points per game}
\description{
NBA data
The data set contains four descriptive statistics for all 105 guards for the 1992–1993 season.
There are many ways to measure the (statistical) performance of guards in the NBA. Of interest is how the
height of the player, minutes per game and free throw percentage affect points per game.
Variable description:
Height: height of the player
MPG: minutes per game
FTP: free throw percentage
PPM: points per game
}
\references{
Chatterjee, S., Handcock, M. S., & Simonoff, J. S. (1995). A casebook for a first course in statistics and data analysis (No. 04; QA276. 12, C4.). New York: Wiley.

Xiang, S., & Yao, W. (2020). Semiparametric mixtures of regressions with single-index for model based clustering. Advances in Data Analysis and Classification, 14(2), 261-292.
}
\keyword{data}
