% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{ROE}
\alias{ROE}
\title{ROE data
contains a total of 2,110 Chinese listed companies on
their Return on Equity (ROE), meaning the amount of net income returned
as a percentage of shareholders equity, which is an important index to measures
a corporation’s profitability. It is also a useful indicator for fundamental
analysts to price the value of stocks.
Variable description:
ROE: Return on Equity}
\description{
ROE data
contains a total of 2,110 Chinese listed companies on
their Return on Equity (ROE), meaning the amount of net income returned
as a percentage of shareholders equity, which is an important index to measures
a corporation’s profitability. It is also a useful indicator for fundamental
analysts to price the value of stocks.
Variable description:
ROE: Return on Equity
}
\references{
Huang, M., Wang, S., Wang, H., & Jin, T. (2018). Maximum smoothed likelihood estimation for a class of semiparametric Pareto mixture densities. Statistics and Its Interface, 11(1), 31-40.
}
\keyword{data}
