% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AFDP}
\alias{AFDP}
\title{AFDP data}
\format{
A data frame containing 7411 observations.
}
\usage{
AFDP
}
\description{
The data contains the 11 sensor measures aggregated over one hour (by means of
average or sum) from a gas turbine located in Turkey's north western region for
the purpose of studying flue gas emissions.
}
\keyword{datasets}
