% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elbow}
\alias{elbow}
\title{Elbow data}
\format{
A data frame containing 507 observations (elbow diameter, sum of two elbows).
}
\usage{
elbow
}
\description{
The data contains elbow dimension measurements on 507 individuals (247 men and
260 women), primarily in their twenties and thirties, all of whom exercise several
hours a week.
}
\references{
Heinz, G., Peterson, L. J., Johnson, R. W., and Kerk, C. J. (2003).
Exploring relationships in body dimensions. Journal of Statistics Education, 11(2)
}
\keyword{datasets}
