% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDEEM.R
\name{KDEEM.H}
\alias{KDEEM.H}
\title{Kernel Density-based EM type algorithm for Semiparametric Mixture Regression with Unspecified Homogenous Error Distributions}
\usage{
KDEEM.H(x, y, M, ini, maxiter)
}
\arguments{
\item{x}{explanatory variables in matrix from with row as observations}

\item{y}{response variable}

\item{M}{number of component}

\item{ini}{initial values for parameters. ini = list(Beta = regression coefficients,
P = matrix: probability of each observation is from component m based on initial beta and H,
H = bandwidth for kernel estimation)}

\item{maxiter}{maximum iteration for the algorithm}
}
\value{
P: posterior probability of each observation is from component m
Beta: estimated regression coefficients
PI: estimated component proportion
H: bandwidth of the kernel estimation
}
\description{
For kernel density-based semiparametric mixture regression for homogeneous error distributions
}
\examples{
# See example in KDEEM
}
