% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mphd.R
\name{mixonekn}
\alias{mixonekn}
\title{Two normal component mixture estimation}
\usage{
mixonekn(x, sig = NULL)
}
\arguments{
\item{x}{observations of two mixture normal}

\item{sig}{standard deviation of first component. By default sig is set to be NULL.}
}
\value{
mu: estimated mean; sigma: estimated sigma; pi:estimated proportion; lh: estimated likelihood for observations.
}
\description{
Estimate the two component mixture model assuming the first component has mean 0.
This function is used in the example of mphd to give initial estimation.
}
\examples{
#see example in mphd
}
