% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA}
\alias{NBA}
\title{NBA data}
\format{
A data frame containing 95 observations and the following 4 variables.
\describe{
\item{Height:}{ height of the player.}
\item{MPG:}{ minutes per game.}
\item{FTP:}{ free throw percentage.}
\item{PPM:}{ points per game.}
}
}
\usage{
NBA
}
\description{
The data contains four descriptive statistics for all 95 guards for the 1992-1993
season. There are many ways to measure the (statistical) performance of guards in
the NBA. Of interest is how the the player's height (Height), minutes per game (MPG),
and free throw percentage (FTP) affect points per game (PPM).
}
\references{
Chatterjee, S., Handcock, M. S., and Simonoff, J. S. (1995). A casebook for a
first course in statistics and data analysis (No. 04; QA276. 12, C4.).
New York: Wiley.

Xiang, S. and Yao, W. (2020). Semiparametric mixtures of regressions with
single-index for model based clustering. Advances in Data Analysis and
Classification, 14(2), 261-292.
}
\keyword{datasets}
