% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROE}
\alias{ROE}
\title{ROE data}
\format{
A data frame containing 2110 observations.
}
\usage{
ROE
}
\description{
The data contains a total of 2110 Chinese listed companies on their Return on
Equity (ROE), which represents the amount of net income returned as a percentage of
shareholders' equity. ROE is an important index used to measure a corporation's
profitability and is also a useful indicator for fundamental analysts to assess
the value of stocks.
}
\references{
Huang, M., Wang, S., Wang, H., and Jin, T. (2018). Maximum smoothed
likelihood estimation for a class of semiparametric Pareto mixture densities.
Statistics and Its Interface, 11(1), 31-40.
}
\keyword{datasets}
