% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM.R
\name{sinvreg}
\alias{sinvreg}
\title{Dimension Reduction Based on Sliced Inverse Regression}
\usage{
sinvreg(x, y, numslice = NULL)
}
\arguments{
\item{x}{independent variable matrix}

\item{y}{response variable vector}

\item{numslice}{number of slice, default is 10}
}
\value{
direction:direction vector
eigenvalue: eigenvalues for reduced x
matx: reduced x
numrep:number of observations in each slice
}
\description{
Use in the examples for sim, simonestep functions
}
\examples{
#see example section of sim function.
}
